+++ tests/applicationinsights_tests/channel_tests/TestSynchronousQueue.py
@@ -4,6 +4,9 @@ import unittest
 import sys
 import os
 import os.path
+
+from tests.applicationinsights_tests.common import MockSynchronousSender
+
 rootDirectory = os.path.join(os.path.dirname(
     os.path.realpath(__file__)), '..', '..')
 if rootDirectory not in sys.path:
@@ -12,11 +15,11 @@ if rootDirectory not in sys.path:
 
 class TestSynchronousQueue(unittest.TestCase):
     def test_construct(self):
-        actual = channel.SynchronousQueue(MockSynchronousSender())
+        actual = channel.SynchronousQueue(MockSynchronousSender(2))
         self.assertIsNotNone(actual)
 
     def test_flush_works_as_expected(self):
-        sender = MockSynchronousSender()
+        sender = MockSynchronousSender(2)
         queue = channel.SynchronousQueue(sender)
         queue.max_queue_length = 3
         for i in range(1, 8):
@@ -26,13 +29,3 @@ class TestSynchronousQueue(unittest.TestCase):
         while queue._queue.qsize() > 0:
             temp.append(queue._queue.get())
         self.assertEqual([7], temp)
-
-
-class MockSynchronousSender:
-    def __init__(self):
-        self.send_buffer_size = 2
-        self.data = []
-        self.queue = None
-
-    def send(self, data_to_send):
-        self.data.append(data_to_send)