+++ tests/applicationinsights_tests/channel_tests/TestSynchronousSender.py
@@ -2,7 +2,6 @@ from applicationinsights import channel
 import unittest
 
 import sys
-import os
 import os.path
 rootDirectory = os.path.join(os.path.dirname(
     os.path.realpath(__file__)), '..', '..')