+++ tests/applicationinsights_tests/channel_tests/contracts_tests/__init__.py
@@ -17,3 +17,5 @@ from .TestOperation import TestOperation
 from .TestSession import TestSession
 from .TestUser import TestUser
 from .TestInternal import TestInternal
+from .TestAvailabilityData import TestAvailabilityData
+from .Utils import TestJsonEncoder