+++ tests/applicationinsights_tests/requests_tests/TestWSGIApplication.py
@@ -1,9 +1,7 @@
 from applicationinsights import requests
 import unittest
-import wsgiref
 
 import sys
-import os
 import os.path
 rootDirectory = os.path.join(os.path.dirname(
     os.path.realpath(__file__)), '..', '..')